backend/requirements.txt
# OpalAI / Panchakarma Agent - Backend
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1

# Gemini / Google AI
google-genai>=1.0.0
Pillow>=11.0.0

# Optional: async HTTP
httpx>=0.28.0
aiofiles>=24.1.0

# CORS
starlette>=0.41.0
